Update Makefile to allow for paths with spaces in them
When I tried to use the Makefile to download the wiki sources it ended up deleting the
directory it was called from, due to the Makefile not being able to handle paths with
spaces in them.

The fix is to wrap the variables that hold the path in quotes when they are being passed
as arguments to commands.